+++ aspnet/mvc/views/forms/sample/final/ViewModels/AddressViewModel.cs
@@ -1,5 +1,6 @@
-﻿
-namespace FormsTH.ViewModels
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace FormsTagHelper.ViewModels
 {
     public class AddressViewModel
     {
@@ -9,7 +10,7 @@ namespace FormsTH.ViewModels
     public class RegisterViewModel2
     {
         public string Email { get; set; }
-
+        [DataType(DataType.Password)]
         public string Password { get; set; }
         public AddressViewModel Address { get; set; }
     }
