+++ aspnet/mvc/views/forms/sample/final/ViewModels/CountryEnum.cs
@@ -1,7 +1,7 @@
 ﻿//#define EnumOnly
 #if EnumOnly
 
-namespace FormsTH.ViewModels
+namespace FormsTagHelper.ViewModels
 {
     public enum CountryEnum
     {
