+++ aspnet/mvc/views/forms/sample/final/ViewModels/CountryEnumDA.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace FormsTH.ViewModels
+namespace FormsTagHelper.ViewModels
 {
     public enum CountryEnum
     {
