+++ aspnet/mvc/views/forms/sample/final/ViewModels/CountryEnumViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace FormsTH.ViewModels
+﻿namespace FormsTagHelper.ViewModels
 {
     public class CountryEnumViewModel
     {
