+++ aspnet/mvc/views/forms/sample/final/ViewModels/CountryEnumViewModelDA.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNet.Mvc.Rendering;
 using System.Collections.Generic;
 
-namespace FormsTH.ViewModels
+namespace FormsTagHelper.ViewModels
 {
     
     public class CountryEnumViewModelDA
