+++ aspnet/mvc/views/forms/sample/final/ViewModels/CountryViewModel.cs
@@ -1,22 +1,22 @@
 ﻿using Microsoft.AspNet.Mvc.Rendering;
 using System.Collections.Generic;
 
-namespace FormsTH.ViewModels
+namespace FormsTagHelper.ViewModels
 {
     public class CountryViewModel
     {
         public string Country { get; set; }
 
-        public List<SelectListItem> Countrys { get; private set; }
+        public List<SelectListItem> Countries { get; private set; }
 
         public CountryViewModel()
         {
 
-            Countrys = new List<SelectListItem>
+            Countries = new List<SelectListItem>
               {
-            new SelectListItem {Value = "MX", Text = "Mexico" },
-            new SelectListItem {Value = "CA", Text = "Canada" },
-            new SelectListItem {Value = "US", Text = "USA"    }
+                  new SelectListItem {Value = "MX", Text = "Mexico" },
+                  new SelectListItem {Value = "CA", Text = "Canada" },
+                  new SelectListItem {Value = "US", Text = "USA"    }
               };
         }
     }
