+++ aspnet/mvc/views/forms/sample/final/ViewModels/RegisterViewModelDisplay.cs
@@ -3,7 +3,7 @@
 
 using System.ComponentModel.DataAnnotations;
 
-namespace FormsTH.ViewModels
+namespace FormsTagHelper.ViewModels
 {
     public class RegisterViewModel
     {
