--- aspnet/mvc/views/forms/sample/final/ViewModels/SomeNumber.cs
@@ -1,11 +1,0 @@
-﻿
-namespace FormsTH.ViewModels
-{
-    public class MyNumber
-    {
-        public decimal SomeNumber { get; set; }
-    }
-
-  
-}
-
