+++ aspnet/mvc/views/forms/sample/final/Views/Demo/Format.cshtml
@@ -1,5 +1,5 @@
-﻿@addTagHelper "*, Microsoft.AspNet.Mvc.TagHelpers"
-@model FormsTH.ViewModels.MyNumber
+﻿
+@model FormsTagHelper.ViewModels.MyNumber
 
 <form asp-controller="DemoReg" asp-action="Format" method="post">
     SomeNumber:  <input asp-for="SomeNumber" asp-format="{0:N4}" /> <br>
