+++ aspnet/mvc/views/forms/sample/final/Views/Demo/Register2.cshtml
@@ -1,5 +1,4 @@
-﻿@model FormsTH.ViewModels.RegisterViewModel2
-@addTagHelper "*, Microsoft.AspNet.Mvc.TagHelpers"
+﻿@model FormsTagHelper.ViewModels.RegisterViewModel2
 
 <form asp-controller="DemoReg" asp-action="Register2" method="post">
     Email:  <input asp-for="Email" /> <br>
