+++ aspnet/mvc/views/forms/sample/final/Views/Demo/RegisterFormOnly.cshtml
@@ -1,7 +1,6 @@
-﻿@model FormsTH.ViewModels.RegisterViewModel
-@addTagHelper "*, Microsoft.AspNet.Mvc.TagHelpers"
+﻿@model FormsTagHelper.ViewModels.RegisterViewModel
 
-<form asp-route="register" method="post">
+<form asp-controller="DemoReg" asp-action="Register" method="post">
     Email:  <input id="Email" name="Email" /> <br>
     Password: <input type="password" id="Password" name="Password" /><br>
     <button type="submit">Register</button>
