+++ aspnet/mvc/views/forms/sample/final/Views/Demo/RegisterInput.cshtml
@@ -1,5 +1,5 @@
-﻿@model FormsTH.ViewModels.RegisterViewModel
-@addTagHelper "*, Microsoft.AspNet.Mvc.TagHelpers"
+﻿@model FormsTagHelper.ViewModels.RegisterViewModel
+
 
 <form asp-controller="DemoReg" asp-action="Register" method="post">
     Email:  <input asp-for="Email" /> <br>
