+++ aspnet/mvc/views/forms/sample/final/Views/Demo/RegisterTA.cshtml
@@ -1,0 +1,7 @@
+﻿@model FormsTagHelper.ViewModels.SimpleViewModel
+
+
+<form asp-controller="DemoReg" asp-action="RegisterTest" method="post">
+    <textarea asp-for="Description"></textarea>
+    <button type="submit">Test</button>
+</form>
