--- aspnet/mvc/views/forms/sample/final/Views/DemoReg/RegisterTA.cshtml
@@ -1,7 +1,0 @@
-﻿@model FormsTH.ViewModels.SimpleViewModel
-@addTagHelper "*, Microsoft.AspNet.Mvc.TagHelpers"
-
-<form asp-controller="DemoReg" asp-action="RegisterTA" method="post">
-    <textarea asp-for="Description"></textarea>
-    <button type="submit">Test</button>
-</form>
