+++ aspnet/mvc/views/forms/sample/final/Views/Home/Index.cshtml
@@ -1,7 +1,7 @@
-﻿@model FormsTH.ViewModels.CountryViewModel
-@addTagHelper "*, Microsoft.AspNet.Mvc.TagHelpers"
+﻿@model FormsTagHelper.ViewModels.CountryViewModel
+
 
 <form asp-controller="Home" asp-action="Index" method="post">
-    <select asp-for="Country" asp-items="Model.Countrys"></select> 
+    <select asp-for="Country" asp-items="Model.Countries"></select> 
     <br /><button type="submit">Register</button>
 </form>
