+++ aspnet/mvc/views/forms/sample/final/Views/Home/IndexEnum.cshtml
@@ -1,7 +1,7 @@
-﻿@using FormsTH.ViewModels
+﻿@using FormsTagHelper.ViewModels
 @model CountryEnumViewModel
 
-@addTagHelper "*, Microsoft.AspNet.Mvc.TagHelpers"
+
 
 <form asp-controller="Home" asp-action="IndexEnum" method="post">
     <select asp-for="EnumCountry"
