+++ aspnet/mvc/views/forms/sample/final/Views/Home/IndexGroup.cshtml
@@ -1,7 +1,7 @@
-﻿@model FormsTH.ViewModels.CountryViewModelGroup
-@addTagHelper "*, Microsoft.AspNet.Mvc.TagHelpers"
+﻿@model FormsTagHelper.ViewModels.CountryViewModelGroup
+
 
 <form asp-controller="Home" asp-action="IndexGroup" method="post">
-    <select asp-for="Country" asp-items="Model.Countrys"></select> 
+    <select asp-for="Country" asp-items="Model.Countries"></select> 
     <br /><button type="submit">Register</button>
 </form>
