+++ aspnet/mvc/views/forms/sample/final/Views/Home/IndexMS.cshtml
@@ -1,7 +1,7 @@
-﻿@model FormsTH.ViewModels.CountryViewModelIEnumerable
-@addTagHelper "*, Microsoft.AspNet.Mvc.TagHelpers"
+﻿@model FormsTagHelper.ViewModels.CountryViewModelIEnumerable
+
 
 <form asp-controller="Home" asp-action="IndexMS" method="post">
-    <select asp-for="CountryCodes" asp-items="Model.Countrys"></select> 
+    <select asp-for="CountryCodes" asp-items="Model.Countries"></select> 
     <br /><button type="submit">Register</button>
 </form>
