+++ aspnet/mvc/views/forms/sample/final/Views/_ViewImports.cshtml
@@ -1,0 +1,1 @@
+﻿@addTagHelper "*, Microsoft.AspNet.Mvc.TagHelpers"
